refactor: migrate folder structure to feature-based architecture
- lib/ui/ → lib/features/ + lib/core/ 구조로 재편
- 각 feature별 독립 모듈화 (book_detail, book_list, reading_start 등)
- 공유 컴포넌트를 core/widgets/, core/view_model/로 이동
- 모든 import 경로 업데이트
