+++ app/lib/core/widgets/book_image_widget.dart
