+++ app/lib/features/book_detail/widgets/book_detail_screen.dart
@@ -19,9 +19,9 @@ import 'package:book_golas/data/services/book_service.dart';
 import 'package:book_golas/data/services/google_vision_ocr_service.dart';
 import 'package:book_golas/data/services/image_cache_manager.dart';
 import 'package:book_golas/domain/models/book.dart';
-import 'package:book_golas/ui/core/ui/book_image_widget.dart';
-import 'package:book_golas/ui/core/ui/custom_snackbar.dart';
-import 'package:book_golas/ui/core/ui/keyboard_accessory_bar.dart';
+import 'package:book_golas/core/widgets/book_image_widget.dart';
+import 'package:book_golas/core/widgets/custom_snackbar.dart';
+import 'package:book_golas/core/widgets/keyboard_accessory_bar.dart';
 import 'dialogs/daily_target_dialog.dart';
 import 'dialogs/today_goal_sheet.dart';
 import 'dialogs/update_page_dialog.dart';
