+++ app/lib/features/book_list/widgets/book_list_screen.dart
@@ -3,8 +3,8 @@ import 'package:shimmer/shimmer.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import 'package:book_golas/domain/models/book.dart';
-import 'package:book_golas/ui/core/ui/book_image_widget.dart';
-import 'book_detail_screen.dart';
+import 'package:book_golas/core/widgets/book_image_widget.dart';
+import 'package:book_golas/features/book_detail/widgets/book_detail_screen.dart';
 
 class BookListScreen extends StatefulWidget {
   const BookListScreen({super.key});
