+++ app/lib/features/reading_start/widgets/reading_start_screen.dart
@@ -6,7 +6,7 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:book_golas/data/services/aladin_api_service.dart';
 import 'package:book_golas/data/services/book_service.dart';
 import 'package:book_golas/domain/models/book.dart';
-import 'package:book_golas/ui/core/ui/book_image_widget.dart';
+import 'package:book_golas/core/widgets/book_image_widget.dart';
 
 class ReadingStartScreen extends StatefulWidget {
   final String? title;
