+++ app/lib/main.dart
@@ -2,16 +2,16 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
-import 'package:book_golas/ui/reading/widgets/reading_chart_screen.dart';
+import 'package:book_golas/features/reading_chart/widgets/reading_chart_screen.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:book_golas/ui/book/widgets/book_list_screen.dart';
-import 'package:book_golas/ui/reading/widgets/reading_start_screen.dart';
+import 'package:book_golas/features/book_list/widgets/book_list_screen.dart';
+import 'package:book_golas/features/reading_start/widgets/reading_start_screen.dart';
 import 'package:book_golas/config/app_config.dart';
 import 'package:book_golas/data/repositories/book_repository.dart';
 import 'package:book_golas/data/services/book_service.dart';
-import 'package:book_golas/ui/home/view_model/home_view_model.dart';
-import 'package:book_golas/ui/core/view_model/theme_view_model.dart';
+import 'package:book_golas/features/home/view_model/home_view_model.dart';
+import 'package:book_golas/core/view_model/theme_view_model.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'firebase_options.dart';
@@ -19,10 +19,10 @@ import 'data/services/auth_service.dart';
 import 'data/services/fcm_service.dart';
 import 'data/services/notification_settings_service.dart';
 import 'data/services/reading_progress_service.dart';
-import 'ui/auth/widgets/login_screen.dart';
-import 'ui/auth/widgets/my_page_screen.dart';
+import 'features/auth/widgets/login_screen.dart';
+import 'features/auth/widgets/my_page_screen.dart';
 import 'domain/models/book.dart';
-import 'ui/book/widgets/book_detail_screen.dart';
+import 'features/book_detail/widgets/book_detail_screen.dart';
 
 // 백그라운드 메시지 핸들러 (main 함수 밖에 정의)
 @pragma('vm:entry-point')
